LayoutTests/media/track/track-webvtt-tc021-valign-expected.txt

Tests cue vertical alignment (direction) from settings.


*** Testing text track 0
EXPECTED (cues.length == '3') OK
EXPECTED (cues[i].direction == 'vertical') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].direction == 'vertical-lr') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].direction == 'vertical') OK
EXPECTED (cues[i].alignment == 'start') OK
EXPECTED (cues[i].textPosition == '0') OK

*** Testing text track 1
EXPECTED (cues.length == '3') OK
EXPECTED (cues[i].direction == 'vertical') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].direction == 'vertical-lr') OK
EXPECTED (cues[i].alignment == 'middle') OK
EXPECTED (cues[i].textPosition == '50') OK
EXPECTED (cues[i].direction == 'vertical') OK
EXPECTED (cues[i].alignment == 'start') OK
EXPECTED (cues[i].textPosition == '0') OK

*** Testing text track 2
EXPECTED (cues.length == '3') OK
EXPECTED (cues[i].direction == 'horizontal') OK
EXPECTED (cues[i].direction == 'horizontal') OK
EXPECTED (cues[i].direction == 'horizontal') OK
END OF TEST
